Introduce hugo generation CI job
We split the generation of the hugo website in a seperate CI job
which will be called from each demo project. Each project
uses the same cache directory. The hugo-generation generates a hugo
website from the data stored in this cache directory. Then it will call
the web-publisher to publish the results in the github pages.

Change-Id: I562a42f28b28f29323e3e11f3927431ac291164f